requirements.txt
Django==3.1.3
beautifulsoup4 == 4.11.1
djangorestframework==3.12.2
Markdown==3.3.7